Delayed expensive castle legality checking until apply function
Sometimes castling will not be examined in search and thus we can save two expensive calls to attacked per call of generatemoves

TODO: look into a scheme where we can look more selectively if we are in check. i.e. only on the diagonal/ortho that the piece moves from (for discover attacks) and for direct attacks by the piece at its new location. (Of course this won't work for king moves as then the target changes and we will have to do the expensive search regardless)